+++ src/main/java/io/github/mhagnumdw/recipes/FormatSqlFile.java
@@ -1,9 +1,10 @@
-package io.github.mhagnumdw;
+package io.github.mhagnumdw.recipes;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.github.vertical_blank.sqlformatter.core.FormatConfig;
 import com.github.vertical_blank.sqlformatter.languages.Dialect;
+import io.github.mhagnumdw.visitors.FormatSqlFileVisitor;
 import lombok.EqualsAndHashCode;
 import lombok.Value;
 import org.jspecify.annotations.Nullable;
@@ -17,12 +18,12 @@ import org.openrewrite.TreeVisitor;
  */
 @Value
 @EqualsAndHashCode(callSuper = false)
-public class FormatSqlFileRecipe extends FormatSqlRecipeAbstract {
+public class FormatSqlFile extends FormatSqlRecipeAbstract {
 
     private static final String DEFAULT_FILE_PATH = "**/*.sql";
 
     @JsonCreator
-    public FormatSqlFileRecipe(@Nullable @JsonProperty("filePath") String filePath,
+    public FormatSqlFile(@Nullable @JsonProperty("filePath") String filePath,
                                @Nullable @JsonProperty("sqlDialect") String sqlDialect,
                                @Nullable @JsonProperty("indent") String indent,
                                @Nullable @JsonProperty("maxColumnLength") Integer maxColumnLength,